refactor: move ledger-protocol to new file
